fixed PYCBC-15; API supports integer value sets
Previously, the developer had to pass strings to .set().
Now, ints can be passed in, and are coerced into strings
both going in and coming out.

WARNING: this is a backwards compatibility break, as old
code which routed around this issue by always expecting
strings to be returned, will now need updating.

Change-Id: I29d215357a15d0c347b07b3a545a76325ef434f7
Reviewed-on: http://review.couchbase.org/18357
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>
Tested-by: Benjamin Young <benjamin@couchbase.com>