put MemcachedClient into its own .py
CouchbaseClient depends on/imports it, of course.
test_memcachedclient.py added, and
test_couchbaseclient.py depends on that.

The reasoning is to allow MemcachedClient to be
tested against a standard memcached server as
well as with a Couchbase Server installation.

nosetest couchbase/tests/test_memcachedclient.py

will run just the memcached tests on the default
11211 port.

Change-Id: Ib5c9b6cc35482dc82d7e7ceff98317ccbcd0932c
Reviewed-on: http://review.couchbase.org/17410
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Benjamin Young <benjamin@couchbase.com>