removed copies of external modules
We will be depending on pip for package installation
for the SDK (as most/all Python packages do). If we
need these dependencies pulled in prior to static
packaging into .zip files, we should look for a
way to do it using the same set of references in
the setup.py (used by pip) and not copy/paste
packages into our repository.

Some backstory:
http://review.couchbase.org/#change,14678

Change-Id: I61d2a342359ccf8b8c6be0e8641a9785ff90220a
Reviewed-on: http://review.couchbase.org/18010
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Benjamin Young <benjamin@couchbase.com>