fixed setup.py version calculation
Previously this was fully dependent on git describe which
meant setup.py had to be run from a source checkout rather
than a source tar ball--which effects pypi distribution as
well as couchbase.com distribution.

This new method uses git describe if available, otherwise it
references a RELEASE-VERSION document stored in the .tar.gz
created by sdist (which happens prior to PyPI uploading).

Change-Id: Ic0bb430532e6e279acb6c1ce6b32a4271c850bd3
Reviewed-on: http://review.couchbase.org/19168
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Benjamin Young <benjamin@couchbase.com>