adding benchmarks testing work
This is based initial work by Pavel which he had
done quickly to generate some initial baseline stats:
https://gist.github.com/1aeaea76519c628702c4

The plan is to test each major operation to get
a sense of how we compare.

python benchmarks/benchmark.py requires that
python-memcached and pylibcb be installed as
they are needed for comparison testing.

Change-Id: I7cca7c5bd9a62a3ad82abe4776eb9f06cd873e23
Reviewed-on: http://review.couchbase.org/19421
Reviewed-by: Benjamin Young <benjamin@couchbase.com>
Tested-by: Benjamin Young <benjamin@couchbase.com>