adding some basic data type handling
The basic code comes from python-memcache.
The compression code was ripped out (for now)
and the focus is on getting strings, ints, longs,
and dictionaries in via the .set() API.

Dicitonaries can now be stored without any additional
json.dumps() code, but end up as strings on output.

This is a known shortcoming, but will not be addressed
until we have some consistent data type handling plan
among the SDKs.

Change-Id: I3b39eee1579ef860d87a5cb94b4bf630a9cfe7d2
Reviewed-on: http://review.couchbase.org/19877
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>
Tested-by: Benjamin Young <benjamin@couchbase.com>