Added get and introducing data formats
You can now also get values via pylibcouchbase.

The default format for setting was changed, it is now set
to "JSON", which means that your Python data gets automatically
encoded as JSON. Other formats are "pickle", which stores the
pickled Python object and there's "plain", which stores the input
as it is given.

It still needs improvements in those areas:
 - error handling
 - being able to set the default format

Tests are currenty missing, but will be added next.