Support for data formats
By default your data will be stored as JSON in Couchbase. You can
also change the format to "pickle" which will use the pickle module
on your object, or "plain" which won't modify your data at all.

Getting the data is transparent, it will do the convesion back to
the original data for you.

This commit also cleans up the code a bit, when using GCC for
compilations there shouldn't be any warnings any more.