More connection options and improved tests
The connection is no longer hard-coded, you can specify the server,
and bucket you want to connect to (together with username and
password).

The test suite was also improved. It's now easier to get it
working with your own Couchbase instance. The parmeters for the
tests are in `tests/tests.ini`.