Remove generated libcouchbase.c
The Cython generated libcouchbase.c resulted in a lot of
conflicts when merging/rebasing commits. It was a pain
to work with, hence it's removed from the source repository.

The generated libcouchbase.c is still be included when
the source gets distributed so that the users don't need
to have Cython installed. Source distributions are
generated via:

    python setup.py sdist

Change-Id: Ib415696b2d80edb337ea1dcb8a9f4c60caba6b0b
Reviewed-on: http://review.couchbase.org/25789
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>