PYCBC-151: Provide backdoor to circumvent abort()
On assertion failures.
An environment variable, PYCBC_ASSERT_CONTINUE
is provided to allow this backdoor option

Rather a noisy message is printed to the screen. This allows
not-so-fatal errors to not hamper application flow. The errors may still
result in a segfault and hence the warning (as the assertions will only
fire when internal state is determined to be inconsistent).

Change-Id: Ided29df3e71b2677bbb1f0eb20ba05f30ccb06b7
Reviewed-on: http://review.couchbase.org/27373
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>