PYCBC-38: Replica read
Note: Due to bugs in the current libcouchbase and/or mock
implementations, the tests fail (and this functionality) does not work
with the upstream libcouchbase version. A new version would be required
to retain proper functionality for the tests

Note that this feature will only be compiled into the extension if
libcouchbase 2.0.7 or greater is used. The feature itself has known
issues in 2.0.7 which will be fixed in the next libcouchbase release.

Change-Id: I1dbd81165450df384c75e2c57a81520e1d741798
Reviewed-on: http://review.couchbase.org/27488
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>