Provide 'ignore_cas' in Item API
This feature already exists in the code. This commit just adds
documentation and a test for the stated feature.

Allow item operations to not be CAS-dependent. By default the existing
'cas' field from the item is used. The 'ignore_cas' parameter instructs
the code to not use the provided CAS for a mutation operation.

Change-Id: Ie4191069207269d2af215620de23dad1bb77e8c7
Reviewed-on: http://review.couchbase.org/27904
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>