Fixup view meta handling
Previously the view meta handling behavior was variable. This commit
makes it so that the view 'meta' will conform to the format like so:

Whatever the format, the various chunks received during the
request shall be appended to an internal list.

Once the complete callback is invoked (from LCB), the extension will
do the following depending on the format type:

FMT_JSON: Will do json.loads("".join(l))
FMT_UTF8: Will do "".join(l)
Anything else, and the list remains as is.

Change-Id: I405ce0fd2af16b4a736618ff21e60b1dd770d0cc