PYCBC-210: Expose LCB timings
Currently this interface is considered private, but it's here - it's
private mainly because of the confusion surrounding how things are
collected.

Timings are enabled via '_start_timings()' and can be collected via
'_get_timings()' - which returns a list of dictionaries containing the
min/max time in milliseconds (or fractions thereof).

Change-Id: I90dfb9b8248ddacd6722e3d0f253a10a3e017b17
Reviewed-on: http://review.couchbase.org/31136
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Volker Mische <volker.mische@gmail.com>