PYCBC-224: Add additional options via '_cntl'
This provides a means by which additional options can be passed to
libcouchbase. The use case here is for options which are not currently
documented within the SDK (either because they have not yet been
exposed, or because a newer libcouchbase library is provided) to be set
without upgrading the client. This is not a supported option but would
provide an alternative to rebuilding the library to have these options
available.

Change-Id: I891949e378f8f75607164badf4daaee3162c03a0
Reviewed-on: http://review.couchbase.org/33698
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>