Remove strict exception equality tests. Use issubclass
With semantic error handling the exceptions thrown might be a subclass
of the expected exception (e.g. NotFoundError). To this end a new helper
method has been implemented in CouchbaseError to check that a specific
return code would end up throwing an exception which is derived from a
specific base class.

Change-Id: Ia950a5206db43b1f4058215b23ad8a97ae2f5007
Reviewed-on: http://review.couchbase.org/37082
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Volker Mische <volker.mische@gmail.com>