Rename gcouchbase and txcouchbase Connection to Bucket
This modifies the naming conventions of the _gevent_ and _Twisted_
modules to be more conforming to the main API.

Additionally, symlinks are added to make editing the files more clear
(it's not helpful in an editor tab to have four files named
"bucket.py").

Finally, in order to facilitate further tests, the old names are now
placed back into `bucket.py`, but are now invoked with a deprecation
warning.

Change-Id: Ibfa9e0277decf6f910b362d73bbe6f5dcb2b6815
Reviewed-on: http://review.couchbase.org/42419
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Volker Mische <volker.mische@gmail.com>