Throw detailed network exceptions upon network errors
This enables the 'detailed_errcodes' functionality of libcouchbase;
available since version 2.4.0. This may modify the behavior a bit in the
client library in the sense that catching 'NetworkError' will no longer
necessarily work (one would need to catch 'CouchbaseNetworkError').

This will make logs a bit easier to read as we'd know exactly what
failed when an operation fails due to network connectivity issues

Change-Id: I16c14558cdf5b8ac9015ef8d62b955a612b05f8c
Reviewed-on: http://review.couchbase.org/44152
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Volker Mische <volker.mische@gmail.com>