Minor fixes to twisted tests so that they'll play nicely with ffi
Unfortunately Python sucks at the proper destruction and memory
management of its objects. Yes, you could write explicit destroy
methods, but that would be a bit counter intuitive and non-idiomatic in
python.

Anyway, this makes the tests a bit more safe when it comes to checking
that the reactor is clean. The test will check if the '_async_shutdown'
method exists (it does in my current cffi impl) and call it. The
'_async_shutdown' is basically a magic close.

Change-Id: Ia3edfba7c12fc051100a6cc2f45b49af4893f55b
Reviewed-on: http://review.couchbase.org/45331
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Volker Mische <volker.mische@gmail.com>