mockserver: set upper limit to wait for initialization
This fixes issues of hanging tests due to missing or corrupt JAR files
(which is sometimes the case when auto-downloading the mock fails).
We rely on the mock connecting to a listening socket (created in the
Python process) - so place a maximum time on waiting for an incoming
connection. This allows the tests not to hang if the mock can't connect
(usually because Java exits "immediately" due to an error).

Change-Id: I683cd42bbef8728423dfaa8069b1b72c5b963602
Reviewed-on: http://review.couchbase.org/48812
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>