Additional view/n1ql documentation fixes
This makes the n1ql methods first class citizens in terms of
documentations. More examples have been added.

The 'itercls' kwarg has been undocumented and made non-discreet, as this
is an internal parameter as far as the API is concerned at the sync
bucket level.

The consistent_with* properties on the N1QLQuery class have been
undocumented as they do not work in Couchbase 4.0 (they might work in a
later release, so they have not been removed).

Change-Id: I999521cbd88909dd9622e32f35d42582f5331d7f
Reviewed-on: http://review.couchbase.org/56411
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>