PYCBC-404, PYCBC-402: GET_COUNT and other subdoc additions
This adds the `get_count()` subdocument operation. It also adds support
for the `get_fulldoc()` and `upsert_fulldoc()` operations which can be
used to retrieve and modify the entire document, respectively.

The get_fulldoc operation requires the use of a not-yet-released
libcouchbase to function properly. The fix is in master.

Change-Id: I75eadfb708beb49c295589f69c08ee364c9023c5
Reviewed-on: http://review.couchbase.org/78996
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>