PYCBC-458: Fix Clang and Python warnings during installation
Motivation
----------
There are various Clang compiler warnings that come up during
installation, which should be addresed. Additionally, the
acouchbase iterator module is included in Python<3.4 packages,
but this module relies on the 'yield from' syntax which is only
available on Python 3.3 and above, and asyncio which is only
available in Python 3.4 and above. We may backport this to
Python2.x using Trollius and Trollius.From in future, but
for now this module should not be included in Python<3.4
packages.

Modifications
-------------
- Fixed various Clang warnings.
- Change Ping code to only fill the 'details' field when an
  error has occurred, matching the behaviour of the JSON
  code from LCB. Also use the svc->rc field rather than
  svc->status, as for LCB. Adjust JSON schema in
  tests to allow for optional 'details' field.
- Selective inclusion of acouchbase.iterator module for
  Python>=3.4
Results
-------
All known warnings/errors eradicated. acouchbase package still
present in Python>=3.4, with all tests passing.

Change-Id: I4ce792fd8e0a1021c2c7fcbff5dfefd9393b5cd0
Reviewed-on: http://review.couchbase.org/89347
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>