PYCBC-489, PYCBC-488: even more rigorous checking for clashing open_bucket params
Motivation
----------

There are three ways of specifying options for open_bucket in the Cluster class.
These are ordered as follows:
1. connstr.options
2. self.authenticator.get_credentials()
3. **kwargs

This provides a great number of ways that mismatching parameters can be
specified. If cert auth parameters are set as well as password parameters,
this presents a conflict between Cert Authentication and password authentication.
Some rules are required to determine whether such potentially risky
ambiguity has occurred.

Changes
-------
Each Authenticator must provide a number of methods to specify:
1. A dictionary of {'fieldname':function} entries containing only the entries
unique to this style of authentication. So, in Cert Auth, this would be
{'certpath': lambda self: self.certpath, 'truststorepath': self.truststorepath}, etc.
2. A set of keys that definitely do not apply to this auth style
3. A method to provide a dictionary of options for bucket authentication
4. A method to provide a dictionary of options for admin authentication

Results
-------

Ambiguous scenarios are logged as warnings, and
Certificate Authentication ambiguities always raise
an exception. Unambiguous scenarios or those
not involving Cert Authentication parameters
continue to work.

Change-Id: I90131b5c7dbf67803222e312049f07ae5694d9e0
Reviewed-on: http://review.couchbase.org/95770
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>