PYCBC-485: sanitize/fix tracing context lifetime management
Motivation
----------
Due to the asynchronous, garbage-collected way Python and
libcouchbase work together, some tracing contexts were never being
completed and their respective spans finished for logging.

This meant that tracing output was missing and also
potentially that memory was being leaked.

Changes
-------
This change adds an 'accounting' process to check that
any child spans of a context that is due to be finished,
based on its reference count. It will clean up these
remaining children if they appear.  If development logging
is enabled, it will list these anomalies. There is
also a strict mode that will immediately stop if
an anomaly is found.

Results
-------
Contexts are always cleaned up for synchronous operations.
For asynchronous operations, it is more difficult to
detect when some spans should end, and hence look
for remaining sub-contexts that require cleaning up.

However, the development logging has flagged up a number
of anomalies already, several of which have been fixed in this
change.

Change-Id: Icc23fff01f8d66ccb621f7ce7af590dfa1ce28d2
Reviewed-on: http://review.couchbase.org/96078
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>