PYCBC-505 - future-proof handling of tag strings
Previous PYCBC code used an optimisation that
made an inference about ownership of the
string returned by lcbtrace_span_get_tag_str when
nstr was zero. This guarantee of ownership
was never explicitly given,
however, and is invalid with the highly
optimised SLLIST-based implementation.

These changes attempt to address that by
copying and zero-terminating the returned string
when necessary.

Change-Id: Ide9e5efdfb5ea00ee3480652824b40cf60bf2689
Reviewed-on: http://review.couchbase.org/96465
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Tested-by: Ellis Breen <ellis.breen@couchbase.com>