PYCBC-499: Admin user_upsert with external type results in error due to password field being sent
MOTIVATION
----------
If using an external authorisation domain, a password is not used or wanted.
We should enforce this by raising an exception in case a password is accidentally provided.

CHANGES
-------
Add an extra check for the presence of a password if using an external auth domain
Also add a check that 'roles' is a non-empty list.

RESULTS
-------
Exception thrown as expected, when passwords are provided with external auth domains.
Exception not thrown when no password is set with an external auth domain.

Change-Id: Ieb9a802f664c762661c0efd8d701f7aeeef61e49
Reviewed-on: http://review.couchbase.org/97501
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>