PYCBC-518: further lost span fixes
MOTIVATION
----------

In some instances, RTO tracing contexts were not being
properly closed because an error code was immediately
returned from the underlying LCB operation, and hence
a callback never returned from which to close the context.

CHANGES
-------

Close contexts in the case of an error
occurring which meant no associated callback would
be triggered, from which the context would normally
be closed.

Various other memory management fixes.

RESULTS
-------

Numerous cases where spans were never closed due
to an error are now handled correctly, so that
contexts are closed and memory returned to the
system.

Change-Id: I70c488241675c2c32abb8572e18181a52eb9e619
Reviewed-on: http://review.couchbase.org/98318
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>