PYCBC-522: lost span fix: minimal set
MOTIVATION
----------
There were some remaining span leaks/lost spans in the
codebase, primarily due to scheduled operations being
forgotten by the client when a failure was
encountered in a multioperation.

CHANGES
-------
Add a 'sched_ops' field to multioperation command
structure to allow us to wait for the completion
and clean up of tracing spans for the successfully
scheduled operations.

RESULTS
-------
Some leaks in some Bucket.get scenarios are now fixed.

Change-Id: I4a0f563e8294489560dd0fa0759820daec589f0d
Reviewed-on: http://review.couchbase.org/99032
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>