PYCBC-523: add parameterized query support for analytics
MOTIVATION
----------

SQL++ for Analytics, like N1QL, supports both positional
and named parameters as part of queries. This should
be added.

CHANGES
-------

Added the relevant arguments to pass up to N1QL,
which expects the same JSON format for arguments.
Added a functional test suite for use against
a real server.

Also renamed couchbase.cbas to couchbase.analytics.
couchbase.cbas is available as an alias for this
release.

RESULTS
-------

Queries with named and positional arguments
return the expected values.

Change-Id: Iac8b575863152f3aec11e77d163eaceafb484674
Reviewed-on: http://review.couchbase.org/99911
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>