PYCBC-493: Add support for host alternate names and ports
MOTIVATION
----------
Couchbase Server and libcouchbase now support alternate
hostnames and ports for use in NAT environments and/or
across virtualised DNS boundaries. We should test
this support.

CHANGES
-------
Added some parameters to the test harness to allow the
specification of the 'network' parameter. Added
tests to check this worked, and ran them on either
side of a Docker install (internal and external) as
well as with the mock (running on localhost).

RESULTS
-------
Support works as expected on all services.

Change-Id: I8aa4de1d2590837733d914c83f3abb2f342216a5
Reviewed-on: http://review.couchbase.org/100896
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>