PYCBC-545: Make JSON sanitization more robust
Motivation
----------
The JSON sanitization code we use in the test
framework to do 'fuzzy' comparisons between JSON output
(from e.g. Analytics) was failing on some Python 2.7
installations due to the lack of a __lt__ method in the
frozendict implementation we use to normalise the
ordering of dictionaries.

Changes
-------
Explicitly sort lists within a JSON object to be
sanitised by their hash value, instead
of relying on the presence of an __lt__method.

Results
-------
Tests work on all tested platforms, against a real
server and the mock.

Change-Id: I700f92e3e80ceed695d0c6f82e2a007aac0a511f
Reviewed-on: http://review.couchbase.org/101848
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>