PYCBC-546: Pull in fix in LCB for test_bad_host segfault
This is needed to pull in the fix for CCBC-866, prior to it being tagged.

Add semver constraints for tags.
These are signified by comma separated terms as follows:

[<>]major.minor.patch(-alpha.[0-9]+|-beta.[0-9]+))
So that when fix is tagged we can use it or later tagged version, and
fall back to last tag entry if nothing else matches (e.g. release-2.10 or master)

Also accepts regex tag constraints due to the limits of CMake semver parsing.

Change-Id: I02d2f26fd02b76d0a1206e0d1e606107efa35690
Reviewed-on: http://review.couchbase.org/109845
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>