PYCBC-593 Restrict LCB versions to tagged >=2.10.3, >=3.0.0-alpha.3
As there may be temporary issues on the HEAD of each branch of LCB as releases are refactored and prepared, enforce that only tagged versions on each given branch greater than or equal to 2.10.3 and 3.0.0-alpha.3 respectively are used.

As this change is on the master (PYCBC 3.x) branch, reduce the list of supported LCBs to just LCB 3.

Change-Id: Id4414ca4a1b790f1005a13572565ae1af2ef60c9
Reviewed-on: http://review.couchbase.org/110824
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>