PYCBC-608: Fix segfault due to deallocated key (release25)
MOTIVATION
----------
When attaching a tracing scope to a given key
prior to scheduling an operation, we decrement the reference count.

This occasionally leads to dangling references to a key
which has been incorrectly destroyed.

The PYCBC_XDECREF command appears to be a remnant from earlier code
which copied the reference and incremented the reference count.

It doesn't appear to be correct or necessary now.

CHANGES
-------
Remove the command to decrement the key reference count.

RESULTS
-------
We no longer experience dangling references to deleted keys.
Code works as expected.

Change-Id: I61d51d691e95bae4bf9e174212d76b0dfcc62dec
Reviewed-on: http://review.couchbase.org/111809
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>