PYCBC-635: Make typing module requirement conditional
Motivation
----------
The Python SDK uses the typing module to provide basic type hints
to users of the library.

This was added in Python 3.5 to the standard library and is available
as an external module for earlier versions of Python.

While this module can be installed on later versions of Python,
this is bad practice as it will clobber the standard library version
for no benefit.

Many build systems may actively prevent this clobbering occurring,
which means that the Python SDK can then not be installed.

Changes
-------
Make the typing module conditional on the version of Python being used.

It will now only install typing if the Python version is less than 3.7.
A similar change was already in master, but this backports it to the 2.x
release line.

Results
-------
For versions of Python 3.7 or greater, the typing module is now no longer
installed.

It is still installed for all releases of Python that need to have it
installed.

Change-Id: Ibf62532191380c199a851c759fe99e1644e752c5
Reviewed-on: http://review.couchbase.org/114013
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>