PYCBC-638: scope and collection methods should be marked as "uncommitted"
Motivation
----------
Scope and collection methods are uncommitted at least until CB Server 6.5
is released. We should mark them as such.

Changes
-------
Start to consolidate warning/annotation functions in one place for
consistency. To this end, renamed 'deprecation.py' to 'supportability.py'.

Add decorator to mark function as uncommitted, which will warn the
user when using the function, and also add a note in the documentation.

This could be used to add deprecation warnings to any Sphinx docs,
later, also. Theoretically, we could annotate deprecated functions
instead of listing them in a class wrapper, later.

Results
-------
Changes work as expected. Python 3 (the support platform) produces a
fully qualified function name for the function in question in its
warning. Python 2 presents the short name.

Change-Id: I3dd3ad9357ba9e36178b99b6dcd97d12bc9c731b
Reviewed-on: http://review.couchbase.org/114474
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>