PYCBC-721 Implement drop_collection and drop_scope
Motivation
=========
They need to be implemented.  While there, it seems there were some
other issues that needed to be dealt with as well.  We were putting
all the collections in the default bucket, for instance.

Modification
============
Corrected some of the paths used in the urls, implemented drop_collection
and drop_scope.  Implemented get_all_scopes as well.  Removed a few
convenience methods that don't appear in the RFC anymore.  NOTE: the
options are in flux (changing the base classes, etc...) so I ignored
the options for now.  Will need to go back and use them when that options
PR is in.

Results
======
New collection manager tests all pass.  They test creation of the scopes
and collections, dropping of them, and get_all_scopes.  Additionally, there
is a test which makes sure the doc you put in a collection really is only
in that collection.

Change-Id: Ic0974048db318a359ddbcbb7342751db1e4f5817
Reviewed-on: http://review.couchbase.org/119628
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>