PYCBC-681 Make OptionBlock arguments discoverable
Motivation
==========
Fancy IDEs (say, PyCharm) can give you a lot of help when constructing
the various Options, if the keyword args are named and typed.  Lets do
that.

Modification
============
So far, did all the existing options and made ones where they didn't
exist for everything in cluster, bucket, and collection.  However,
we still need DiagnosticsOptions in cluster, but the function also needs to
be fixed (seems to hang forever).  So lets fix this all in PYCBC-792.
Also bucket.ping needs to be done in PYCBC-723 - that is waiting on a
small LCB change.

Results
=======
Existing tests all pass, some of this is exercised but not all.  As
the v2 tests get converted, the options will be called pretty thoroughly,
which should cover things fairly well

Change-Id: I429d3044149f26b923defbd95d477ef4fe16861f
Reviewed-on: http://review.couchbase.org/121374
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>